requirements.txt
html2text==2025.4.15
pillow==11.3.0
opencv-python
torch
torchvision
transformers
qdrant-client>=1.17.0
sentencepiece
openai>=1.0.0
tqdm
ultralytics>=8.0.0
supervision>=0.20.0